Factored out FailedConstraint
