+++ collect_app/src/main/java/org/odk/collect/android/javarosawrapper/FailedConstraint.java
@@ -1,0 +1,13 @@
+package org.odk.collect.android.javarosawrapper;
+
+import org.javarosa.core.model.FormIndex;
+
+public class FailedConstraint {
+    public final FormIndex index;
+    public final int status;
+
+    public FailedConstraint(FormIndex index, int status) {
+        this.index = index;
+        this.status = status;
+    }
+}
